==> notebooks/01_02.ipynb <==
# NLP Basics: What Is Natural Language Processing & The Natural Language Toolkit?

### How To Install NLTK On Your Local Machine

Both sets of instructions below assume you already have Python installed. These instructions are taken directly from [http://www.nltk.org/install.html](http://www.nltk.org/install.html).

**Mac/Unix**

From the terminal:
1. Install NLTK: run `pip install -U nltk`
2. Test installation: run `python` then type `import nltk`

**Windows**

1. Install NLTK: [http://pypi.python.org/pypi/nltk](http://pypi.python.org/pypi/nltk)
2. Test installation: `Start>Python`, then type `import nltk`

### Download NLTK Data

# Import the package and download the necessary data
import nltk
nltk.download()

# What methods and attributes are available in nltk?

### What Can You Do With NLTK?

# Did it work?
from nltk.corpus import stopwords

stopwords.words('english')[0:5]

# Let's look at additional words later in the list
